Rename module to override_authority
