+++ linkerd/app/outbound/src/lib.rs
@@ -157,7 +157,7 @@ impl Config {
                     }))
                     .push(observability.clone())
                     .push(identity_headers.clone())
-                    .push(http::overwrite_authority::Layer::new(vec![HOST.as_str(), CANONICAL_DST_HEADER]))
+                    .push(http::override_authority::Layer::new(vec![HOST.as_str(), CANONICAL_DST_HEADER]))
                     // Ensures that the request's URI is in the proper form.
                     .push(http::normalize_uri::layer())
                     // Upgrades HTTP/1 requests to be transported over HTTP/2 connections.
